feat(cmd): merge source and tag into one command
