+++ .github/workflows/rust.yml
@@ -30,5 +30,5 @@ jobs:
           cargo build -vv
       - name: Upgrade template
         run: |
-          ./target/release/sup update
+          ./target/release/sup --update
           ./target/release/sup switch -t v2.0.0-rc6
